# PERF: DataFrame.clip / Series.clip