--- a/asv_bench/benchmarks/frame_methods.py
+++ b/asv_bench/benchmarks/frame_methods.py
@@ -17,6 +17,24 @@ from pandas import (
 from .pandas_vb_common import tm
 
 
+class Clip:
+    params = [
+        ["float64", "Float64", "float64[pyarrow]"],
+        [True, False],
+    ]
+    param_names = ["dtype", "hasna"]
+
+    def setup(self, dtype, hasna):
+        data = np.random.randn(100_000, 10)
+        df = DataFrame(data, dtype=dtype)
+        if hasna:
+            df.iloc[2, :] = None
+        self.df = df
+
+    def time_clip(self, dtype, hasna):
+        self.df.clip(-1.0, 1.0)
+
+
 class GetNumericData:
     def setup(self):
         self.df = DataFrame(np.random.randn(10000, 25))