--- a/pandas/tests/copy_view/test_clip.py
+++ b/pandas/tests/copy_view/test_clip.py
@@ -12,10 +12,10 @@ def test_clip_inplace_reference(using_copy_on_write):
     view = df[:]
     df.clip(lower=2, inplace=True)
 
-    # Clip not actually inplace right now but could be
-    assert not np.shares_memory(get_array(df, "a"), arr_a)
-
     if using_copy_on_write:
+        assert not np.shares_memory(get_array(df, "a"), arr_a)
         assert df._mgr._has_no_reference(0)
         assert view._mgr._has_no_reference(0)
         tm.assert_frame_equal(df_copy, view)
+    else:
+        assert np.shares_memory(get_array(df, "a"), arr_a)