--- a/pandas/tests/copy_view/test_clip.py
+++ b/pandas/tests/copy_view/test_clip.py
@@ -28,13 +28,12 @@ def test_clip_inplace_reference_no_op(using_copy_on_write):
     view = df[:]
     df.clip(lower=0, inplace=True)
 
+    assert np.shares_memory(get_array(df, "a"), arr_a)
+
     if using_copy_on_write:
-        assert np.shares_memory(get_array(df, "a"), arr_a)
         assert not df._mgr._has_no_reference(0)
         assert not view._mgr._has_no_reference(0)
         tm.assert_frame_equal(df_copy, view)
-    else:
-        assert not np.shares_memory(get_array(df, "a"), arr_a)
 
 
 def test_clip_inplace(using_copy_on_write):
@@ -42,8 +41,7 @@ def test_clip_inplace(using_copy_on_write):
     arr_a = get_array(df, "a")
     df.clip(lower=2, inplace=True)
 
-    # Clip not actually inplace right now but could be
-    assert not np.shares_memory(get_array(df, "a"), arr_a)
+    assert np.shares_memory(get_array(df, "a"), arr_a)
 
     if using_copy_on_write:
         assert df._mgr._has_no_reference(0)