@lukemanley 2023-02-19
Thanks for the explanation. I've updated the PR to use `BlockManager.where` and added a test for the int/float case you highlighted. I've updated the timings the OP.