@phofl 2023-02-19
You could look into put mask for inplace ops