@phofl 2023-02-20
Merged the clip tests, could you rebase?


Ah that makes sense, got it.